Added error checking to auth functions

+++ JobHunta/app/watchlist.py
@@ -36,17 +36,23 @@ def add_to_watchlist(u_id, job_posting):
 
     job_id = job_posting['url']
 
+    # Checking if job already exists
+    cur.execute("SELECT * FROM job WHERE id = ?;", job_id)
 
-    job_data = (job_posting['url'],
-                job_posting['title'],
-                job_posting['job_type'],
-                job_posting['description'],
-                job_posting['location'],
-                job_posting['company'],
-                job_posting['created'],
-                job_posting['salary'])
+    # Adding to job
+    if cur.rowcount == 0:
+        job_data = (job_posting['url'],
+                    job_posting['title'],
+                    job_posting['job_type'],
+                    job_posting['description'],
+                    job_posting['location'],
+                    job_posting['company'],
+                    job_posting['created'],
+                    job_posting['salary'])
 
-    cur.execute("INSERT INTO job VALUES (?, ?, ?, ?, ?, ?, ?, ?) ;", job_data)
+        cur.execute("INSERT INTO job VALUES (?, ?, ?, ?, ?, ?, ?, ?) ;", job_data)
+
+    # Adding to watchlist
     cur.execute("INSERT INTO watchlist VALUES (?, ?);", (u_id, job_id))
 
     conn.commit()
@@ -59,7 +65,7 @@ def in_watchlist(u_id, job_id):
     conn = db.get_db()
     cur = conn.cursor()
 
-    cur.execute("SELECT * FROM watchlist WHERE u_id = ? AND job_id = ?;", (u_id, job_id))
+    cur.execute("SELECT * FROM watchlist WHERE user_id = ? AND job_id = ?;", (u_id, job_id))
 
     result = cur.rowcount > 0
 
@@ -68,15 +74,17 @@ def in_watchlist(u_id, job_id):
     return result
 
 # Removes job posting to watchlist
-def remove_from_watchlist(u_id, url):
+def remove_from_watchlist(u_id, job_id):
     # Getting db and cursor
     conn = db.get_db()
     cur = conn.cursor()
 
-    job_id = url
-    cur.execute("DELETE FROM watchlist WHERE user_id = ? AND job_id = ?;", (u_id, job_id))
+    cur.execute("SELECT * FROM watchlist WHERE user_id = ? AND job_id = ?;", (u_id, job_id))
 
-    conn.commit()
+    if cur.rowcount != 0:
+        cur.execute("DELETE FROM watchlist WHERE user_id = ? AND job_id = ?;", (u_id, job_id))
+        conn.commit()
+
     db.close_db()
 
     return True
